backend/requirements.txt
Django==2.0
djangorestframework==3.7.7
mysqlclient==1.3.12
redis==2.10.6
sse==1.2
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-cors-headers==2.2.0
Werkzeug==0.14.1
ipython==6.2.1
git+https://github.com/niwinz/django-sse.git
-e git+https://github.com/dongweiming/wxpy@signals#egg=wxpy
-e git+https://github.com/dongweiming/ItChat@signals#egg=itchat